chores: Change to Spring Data JPA and implement automatic auditing supports

+++ src/main/java/com/quentin/eazyschool/payload/ContactDTO.java
@@ -5,7 +5,8 @@ import lombok.Data;
 
 @Data
 public class ContactDTO {
-    private int contactId;
+
+    private Long contactId;
 
     @NotBlank(message = "Name must not be blank")
     private String name;
